requirements.txt
PyMuPDF==1.22.4
yara-python==4.3.1
configparser==5.3.0  # For handling the configuration file

